+++ js/util.js
@@ -1,0 +1,13 @@
+// Функция, возвращающая случайное число из диапазона
+
+const getRandomNumber = (a, b) => {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export {getRandomNumber};
